Display error message if project can't be loaded because of nonexistence or permissions issues
