+++ common/components/controllers/AboutProjectController.jsx
@@ -21,10 +21,12 @@ import type {PositionInfo} from "../forms/PositionInfo.jsx";
 import Headers from "../common/Headers.jsx";
 import Truncate from "../utils/truncate.js";
 import IconLinkDisplay from "../componentsBySection/AboutProject/IconLinkDisplay.jsx";
+import {APIError} from "../utils/api.js";
 
 
 type State = {|
   project: ?ProjectDetailsAPIData,
+  loadStatusMsg: string,
   showJoinModal: boolean,
   positionToJoin: ?PositionInfo,
   showPositionModal: boolean,
@@ -38,6 +40,7 @@ class AboutProjectController extends React.PureComponent<{||}, State> {
     super();
     this.state = {
     project: null,
+    loadStatusMsg: "Loading...",
     showContactModal: false,
     showPositionModal: false,
     shownPosition: null,
@@ -50,7 +53,7 @@ class AboutProjectController extends React.PureComponent<{||}, State> {
 
   componentDidMount() {
     const projectId: string = (new RegExp("id=([^&]+)")).exec(document.location.search)[1];
-    ProjectAPIUtils.fetchProjectDetails(projectId, this.loadProjectDetails.bind(this));
+    ProjectAPIUtils.fetchProjectDetails(projectId, this.loadProjectDetails.bind(this), this.handleLoadProjectFailure.bind(this));
     metrics.logNavigateToProjectProfile(projectId);
   }
 
@@ -59,7 +62,14 @@ class AboutProjectController extends React.PureComponent<{||}, State> {
       project: project,
     });
   }
-
+  
+  handleLoadProjectFailure(error: APIError) {
+    this.setState({
+      loadStatusMsg: "Could not load project"
+    });
+  }
+  
+  
   handleShowVolunteerModal(position: ?PositionInfo) {
     this.setState({
       showJoinModal: true,
@@ -87,7 +97,7 @@ class AboutProjectController extends React.PureComponent<{||}, State> {
   }
 
   render(): $React$Node {
-    return this.state.project ? this._renderDetails() : <div>Loading...</div>
+    return this.state.project ? this._renderDetails() : <div>{this.state.loadStatusMsg}</div>
   }
 
   _renderDetails(): React$Node {
