+++ common/components/utils/api.js
@@ -1,0 +1,10 @@
+// @flow
+
+export type APIResponse = {|
+  +status: number
+|};
+
+export type APIError = {|
+  +errorCode: number,
+  +errorMessage: string
+|};
